requirements.txt
Django>=3.2
djangorestframework>=3.14
psycopg2-binary>=2.9
django-environ>=0.10
celery>=5.3
redis>=4.5
django.contrib.gis
python-telegram-bot>=13.7
requests>=2.25 